Password Indexer
My goal in this project is to hash the passwords in text files that may contain millions of passwords according to MD5, SHA1, SHA256 formats, open a folder according to the first character of each password and write the relevant password to the text file in the relevant folder. In short, this program indexes and groups passwords. Apart from these, there is also a password search section and if a password entered here is among these millions of passwords, it finds it and notifies the user, if not, it saves this newly written password to the relevant Index folder. It took a long time to finalize this code and after finalizing it and making some optimizations, a program emerged that indexed millions of passwords in about 20-30 seconds. The first working version of the program took 4-5 minutes to complete the indexing process. From this perspective, I made a good optimization. Later, I realized the mistakes I made and shortened this time by making changes to the indexing system. As a result, this project gave me good experience in file organization and working with files. Apart from these, it was also a good experience in terms of the optimization problems I fixed.